Removed OpenBots.UI
