+++ OpenBots.Commands/OpenBots.Commands.ErrorHandling/OpenBots.Commands.ErrorHandling.csproj
@@ -65,11 +65,6 @@
       <Project>{6956836d-d2b0-4b7a-b06d-fa3a22b4c4ff}</Project>
       <Name>OpenBots.Core</Name>
     </ProjectReference>
-    
-    <ProjectReference Include="..\..\OpenBots.UI\OpenBots.UI.csproj">
-      <Project>{35faa5da-fe90-4398-a1ed-2d89f6f2f7b3}</Project>
-      <Name>OpenBots.UI</Name>
-    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
