+++ OpenBots.Commands/OpenBots.Commands.Image/SurfaceAutomationCommand.cs
@@ -9,8 +9,6 @@ using OpenBots.Core.User32;
 using OpenBots.Core.Utilities.CommandUtilities;
 using OpenBots.Core.Utilities.CommonUtilities;
 using OpenBots.Core.Utilities.FormsUtilities;
-
-using OpenBots.UI.Utilities;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -39,6 +37,7 @@ namespace OpenBots.Commands.Image
 				 "For images that are primarily white space, tagging color to the top-left corner of the image and setting \n" +
 				 "the relative click position will produce faster results.")]
 		[Editor("ShowImageCaptureHelper", typeof(UIAdditionalHelperType))]
+		[Editor("ShowImageRecognitionTestHelper", typeof(UIAdditionalHelperType))]
 		public string v_ImageCapture { get; set; }
 
 		[Required]
@@ -131,12 +130,6 @@ namespace OpenBots.Commands.Image
 				throw new InvalidDataException("Accuracy value is invalid");
 			}
 
-			if (testMode)
-			{
-				UICommandsHelper.FindImageElement(userImage, accuracy, TestMode);
-				return;
-			}
-
 			dynamic element = null;
 			if (v_ImageAction == "Wait For Image To Exist")
 			{
@@ -152,7 +145,7 @@ namespace OpenBots.Commands.Image
 				{
 					try
 					{
-						element = UICommandsHelper.FindImageElement(userImage, accuracy, TestMode);
+						element = CommandsHelper.FindImageElement(userImage, accuracy);
 
 						if (element == null)
 							throw new Exception("Image Element Not Found");
@@ -172,7 +165,7 @@ namespace OpenBots.Commands.Image
 				return;
 			}
 			else
-				element = UICommandsHelper.FindImageElement(userImage, accuracy, TestMode);
+				element = CommandsHelper.FindImageElement(userImage, accuracy);
 
 			try
 			{
