+++ OpenBots.Studio/UI/CustomControls/CommandControls.cs
@@ -10,13 +10,14 @@ using OpenBots.Core.Settings;
 using OpenBots.Core.UI.Controls;
 using OpenBots.Core.UI.Controls.CustomControls;
 using OpenBots.Core.Utilities.CommandUtilities;
+using OpenBots.Core.Utilities.CommonUtilities;
+using OpenBots.Core.Utilities.FormsUtilities;
 using OpenBots.Engine;
 using OpenBots.Studio.Utilities;
 using OpenBots.UI.Forms;
 using OpenBots.UI.Forms.ScriptBuilder_Forms;
 using OpenBots.UI.Forms.Supplement_Forms;
 using OpenBots.UI.SupplementForms;
-using OpenBots.UI.Utilities;
 using Serilog.Core;
 using System;
 using System.Collections.Generic;
@@ -24,6 +25,7 @@ using System.ComponentModel;
 using System.Data;
 using System.Diagnostics;
 using System.Drawing;
+using System.Drawing.Imaging;
 using System.Linq;
 using System.Reflection;
 using System.Threading;
@@ -440,18 +442,12 @@ namespace OpenBots.UI.CustomControls
                         helperControl.CommandImage = Resources.command_camera;
                         helperControl.CommandDisplay = "Capture Reference Image";
                         helperControl.Click += (sender, e) => ShowImageCapture(sender, e);
+                        break;
 
-                        CommandItemControl testRun = new CommandItemControl();
-                        testRun.Padding = new Padding(10, 0, 0, 0);
-                        testRun.ForeColor = Color.AliceBlue;
-
-                        testRun.CommandImage = Resources.command_camera;
-                        testRun.CommandDisplay = "Run Image Recognition Test";
-                        testRun.ForeColor = Color.AliceBlue;
-                        testRun.Font = new Font("Segoe UI Semilight", 10);
-                        testRun.Tag = targetControls.FirstOrDefault();
-                        testRun.Click += (sender, e) => RunImageCapture(sender, e);
-                        controlList.Add(testRun);
+                    case UIAdditionalHelperType.ShowImageRecognitionTestHelper:
+                        helperControl.CommandImage = Resources.command_camera;
+                        helperControl.CommandDisplay = "Run Image Recognition Test";
+                        helperControl.Click += (sender, e) => RunImageCapture(sender, e);
                         break;
 
                     case UIAdditionalHelperType.ShowCodeBuilder:
@@ -764,7 +760,9 @@ namespace OpenBots.UI.CustomControls
 
             try
             {
-                UICommandsHelper.FindImageElement(new Bitmap(Common.Base64ToImage(imageSource)), 0.8, true);
+                ImageElement element = FindImageElementTest(new Bitmap(Common.Base64ToImage(imageSource)), 0.8);
+                if (element == null)
+                    MessageBox.Show("Image not found");
             }
             catch (Exception ex)
             {
@@ -772,6 +770,137 @@ namespace OpenBots.UI.CustomControls
             }
             //show all forms
             ShowAllForms();
+        }
+
+        public ImageElement FindImageElementTest(Bitmap smallBmp, double accuracy)
+        {
+            FormsHelper.HideAllForms();
+
+            dynamic element = null;
+            double tolerance = 1.0 - accuracy;
+
+            Bitmap bigBmp = ImageMethods.Screenshot();
+
+            Bitmap smallTestBmp = new Bitmap(smallBmp);
+
+            Bitmap bigTestBmp = new Bitmap(bigBmp);
+            Graphics bigTestGraphics = Graphics.FromImage(bigTestBmp);
+
+            BitmapData smallData =
+              smallBmp.LockBits(new Rectangle(0, 0, smallBmp.Width, smallBmp.Height),
+                       ImageLockMode.ReadOnly,
+                       PixelFormat.Format24bppRgb);
+            BitmapData bigData =
+              bigBmp.LockBits(new Rectangle(0, 0, bigBmp.Width, bigBmp.Height),
+                       ImageLockMode.ReadOnly,
+                       PixelFormat.Format24bppRgb);
+
+            int smallStride = smallData.Stride;
+            int bigStride = bigData.Stride;
+
+            int bigWidth = bigBmp.Width;
+            int bigHeight = bigBmp.Height - smallBmp.Height + 1;
+            int smallWidth = smallBmp.Width * 3;
+            int smallHeight = smallBmp.Height;
+
+            int margin = Convert.ToInt32(255.0 * tolerance);
+
+            unsafe
+            {
+                byte* pSmall = (byte*)(void*)smallData.Scan0;
+                byte* pBig = (byte*)(void*)bigData.Scan0;
+
+                int smallOffset = smallStride - smallBmp.Width * 3;
+                int bigOffset = bigStride - bigBmp.Width * 3;
+
+                bool matchFound = true;
+
+                for (int y = 0; y < bigHeight; y++)
+                {
+                    for (int x = 0; x < bigWidth; x++)
+                    {
+                        byte* pBigBackup = pBig;
+                        byte* pSmallBackup = pSmall;
+
+                        //Look for the small picture.
+                        for (int i = 0; i < smallHeight; i++)
+                        {
+                            int j = 0;
+                            matchFound = true;
+                            for (j = 0; j < smallWidth; j++)
+                            {
+                                //With tolerance: pSmall value should be between margins.
+                                int inf = pBig[0] - margin;
+                                int sup = pBig[0] + margin;
+                                if (sup < pSmall[0] || inf > pSmall[0])
+                                {
+                                    matchFound = false;
+                                    break;
+                                }
+
+                                pBig++;
+                                pSmall++;
+                            }
+
+                            if (!matchFound)
+                                break;
+
+                            //We restore the pointers.
+                            pSmall = pSmallBackup;
+                            pBig = pBigBackup;
+
+                            //Next rows of the small and big pictures.
+                            pSmall += smallStride * (1 + i);
+                            pBig += bigStride * (1 + i);
+                        }
+
+                        //If match found, we return.
+                        if (matchFound)
+                        {
+                            element = new ImageElement
+                            {
+                                LeftX = x,
+                                MiddleX = x + smallBmp.Width / 2,
+                                RightX = x + smallBmp.Width,
+                                TopY = y,
+                                MiddleY = y + smallBmp.Height / 2,
+                                BottomY = y + smallBmp.Height
+                            };
+
+                            //draw on output to demonstrate finding
+                            var Rectangle = new Rectangle(x, y, smallBmp.Width - 1, smallBmp.Height - 1);
+                            Pen pen = new Pen(Color.Red);
+                            pen.Width = 5.0F;
+                            bigTestGraphics.DrawRectangle(pen, Rectangle);
+
+                            frmImageCapture captureOutput = new frmImageCapture();
+                            captureOutput.pbTaggedImage.Image = smallTestBmp;
+                            captureOutput.pbSearchResult.Image = bigTestBmp;
+                            captureOutput.TopMost = true;
+                            captureOutput.Show();
+                            
+                            break;
+                        }
+                        //If no match found, we restore the pointers and continue.
+                        else
+                        {
+                            pBig = pBigBackup;
+                            pSmall = pSmallBackup;
+                            pBig += 3;
+                        }
+                    }
+
+                    if (matchFound)
+                        break;
+
+                    pBig += bigOffset;
+                }
+            }
+
+            bigBmp.UnlockBits(bigData);
+            smallBmp.UnlockBits(smallData);
+            bigTestGraphics.Dispose();
+            return element;
         }
 
         public Tuple<string, string> ShowConditionElementRecorder(object sender, EventArgs e, IfrmCommandEditor editor)
