+++ OpenBots.Studio/UI/Forms/Supplement Forms/frmImageCapture.resx
