--- OpenBots.UI/OpenBots.UI.nuspec
@@ -1,27 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<package>
-  <metadata>
-    <id>OpenBots.UI</id>
-    <version>1.2.1.0</version>
-    <title>OpenBots UI</title>
-    <authors>OpenBots Inc</authors>
-    <owners>OpenBots Inc</owners>
-    <licenseUrl>https://github.com/OpenBotsAI/OpenBots.Studio/blob/develop/LICENSE</licenseUrl>
-    <projectUrl>https://openbots.ai</projectUrl>
-    <iconUrl>https://openbots.ai/wp-content/uploads/2020/11/Studio-Icon-256px.png</iconUrl>
-    <repository type="git" url="https://github.com/OpenBotsAI/OpenBots.Studio" branch="master" />
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>OpenBots UI library for interacting with Commands and Studio.</description>
-    <releaseNotes>https://openbots.ai/category/release-notes/</releaseNotes>
-    <copyright>(C) OpenBots Inc 2020-2021</copyright>
-    <tags>OpenBots Core Studio Command</tags>
-    <dependencies>
-      <group targetFramework="net48">
-      </group>
-    </dependencies>
-  </metadata>
-  <files>
-    <file src="bin\Debug\*.dll" target="lib\net48" />
-		<file src="bin\Debug\*.pdb" target="lib\net48" />
-  </files>
-</package>
