+++ OpenBots.sln
@@ -65,8 +65,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenBots.Commands.QueueItem
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenBots.Commands.IEBrowser", "OpenBots.Commands\OpenBots.Commands.IEBrowser\OpenBots.Commands.IEBrowser.csproj", "{FC1E759F-CD70-44CE-BE76-5AF54222E0FC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenBots.UI", "OpenBots.UI\OpenBots.UI.csproj", "{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenBots.Commands.If", "OpenBots.Commands\OpenBots.Commands.If\OpenBots.Commands.If.csproj", "{D104C957-9DC9-4F97-8BCE-00B21720EF2F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenBots.Commands.ErrorHandling", "OpenBots.Commands\OpenBots.Commands.ErrorHandling\OpenBots.Commands.ErrorHandling.csproj", "{BA2B874F-7389-441F-BDE2-60ADBFE31C9D}"
@@ -469,18 +467,6 @@ Global
 		{FC1E759F-CD70-44CE-BE76-5AF54222E0FC}.Release|x64.Build.0 = Release|Any CPU
 		{FC1E759F-CD70-44CE-BE76-5AF54222E0FC}.Release|x86.ActiveCfg = Release|Any CPU
 		{FC1E759F-CD70-44CE-BE76-5AF54222E0FC}.Release|x86.Build.0 = Release|Any CPU
-		{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}.Debug|x64.Build.0 = Debug|Any CPU
-		{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}.Debug|x86.Build.0 = Debug|Any CPU
-		{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}.Release|x64.ActiveCfg = Release|Any CPU
-		{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}.Release|x64.Build.0 = Release|Any CPU
-		{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}.Release|x86.ActiveCfg = Release|Any CPU
-		{35FAA5DA-FE90-4398-A1ED-2D89F6F2F7B3}.Release|x86.Build.0 = Release|Any CPU
 		{D104C957-9DC9-4F97-8BCE-00B21720EF2F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D104C957-9DC9-4F97-8BCE-00B21720EF2F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D104C957-9DC9-4F97-8BCE-00B21720EF2F}.Debug|x64.ActiveCfg = Debug|Any CPU
